+++ src/DataProtection/DataProtection.Autorest/custom/Cmdlets/Platform/Restore/Initialize-AzDataProtectionRestoreRequest.ps1
@@ -113,6 +113,10 @@ function Initialize-AzDataProtectionRestoreRequest
         [Hashtable]
         ${PrefixMatch},
 
+        [Parameter(ParameterSetName="AlternateLocationILR", Mandatory=$false, HelpMessage='Use this parameter to rename container(s) for alternate location ILR. Input for this parameter is a hashtable where each key is the original container name and each value is the new name for the corresponding container.')]
+        [Hashtable]
+        ${RenameTo},
+
         [Parameter(ParameterSetName="OriginalLocationILR", Mandatory=$false, HelpMessage='Specify the blob restore start range for PITR. You can use this option to specify the starting range for a subset of blobs in each container to restore. use a forward slash (/) to separate the container name from the blob prefix pattern.')]
         # [Parameter(ParameterSetName="AlternateLocationILR", Mandatory=$false, HelpMessage='Minimum matching value for Item Level Recovery.')]
         [System.String[]]
@@ -296,12 +300,18 @@ function Initialize-AzDataProtectionRestoreRequest
             $restoreRequest.RestoreTargetInfo.ObjectType = "itemLevelRestoreTargetInfo"
 
             $restoreCriteriaList = @()
+
+            # Validate -RenameTo against the manifest before branching on datasource type so unsupported workloads fail fast.
+            if($PSBoundParameters.ContainsKey("RenameTo") -and ($manifest.renameContainersEnabled -ne $true)){
+                throw "DatasourceType $DatasourceType does not support renaming containers"
+            }
             
             # can generalise this condition to manifest level if needed
             if($DatasourceType -ne "AzureKubernetesService"){ # TODO: remove Datasource dependency
                 
                 if(($RecoveryPoint -ne $null) -and ($RecoveryPoint -ne "") -and $ContainersList.length -gt 0){
                     $hasPrefixMatch = $PSBoundParameters.Remove("PrefixMatch")
+                    $hasRenameTo = $PSBoundParameters.Remove("RenameTo")
                     for($i = 0; $i -lt $ContainersList.length; $i++){
                                 
                         $restoreCriteria = [Microsoft.Azure.PowerShell.Cmdlets.DataProtection.Models.ItemPathBasedRestoreCriteria]::new()
@@ -316,6 +326,14 @@ function Initialize-AzDataProtectionRestoreRequest
                                 throw "values for PrefixMatch must be string array for each container"
                             }
                             $restoreCriteria.SubItemPathPrefix = $pathPrefix
+                        }
+
+                        if($manifest.renameContainersEnabled -eq $true -and $hasRenameTo){
+                            $renameToValue = $RenameTo[$ContainersList[$i]]
+                            if($renameToValue -ne $null -and ($renameToValue -is [Array])){
+                                throw "value for RenameTo must be a string for each container"
+                            }
+                            $restoreCriteria.RenameTo = $renameToValue
                         }
 
                         # adding a criteria for each container given