+++ src/Network/Network/ChangeLog.md
@@ -20,7 +20,7 @@
 
 ## Upcoming Release
 * Fixed `Approve-AzPrivateEndpointConnection`, `Deny-AzPrivateEndpointConnection`, `Get-AzPrivateEndpointConnection`, `Remove-AzPrivateEndpointConnection`, and `Get-AzPrivateLinkResource` to properly inherit dynamic parameters from base cmdlet class
-* Updated `UserAssignedIdentityId` parameter type from string to string[] in Azure Firewall Policy to input multiple UAMIs
+* Changed `UserAssignedIdentityId` type from string to string[] in Azure Firewall Policy
 
 ## Version 7.27.0
 * Added support to associate a DDoS custom policy with a Load Balancer frontend IP configuration.